try restrictive name matchings if too many download candidates. Fix #3
Download type (name) specified via --type was just matched as a substring of the
download candidates. In some cases this was not enough to narrow down candidates
to one, even after applying the other filters and preference rules.

For example, for candidates 'Download' and 'Download HD' it was impossible to
select the first. Using --type 'Download' would match both, and no criteria
could possibly exclude just the HD one.

Also, for candidates '.bin (wine)' and 'Native .bin', a type '.bin' should have
a preference for the first candidate, as it starts with the criteria while the
other merely contains it.

The solution in this commit is:

When there are still multiple candidates after applying all filters and
preference rules, see which candidates starts with the criteria, and if just
one matches, choose it. This would solve the second example.

If still multiple candidates, try matching those that *are* the criteria. This
This would solve the first example.

All three matching rules ("contains", "starts with", "is") are case-insensitive,
as I doubt a case-sensitive matching would be of any help.